+++ src/observability/api.rs
@@ -1124,7 +1124,7 @@ impl ObservabilityApi {
                         memory_total_mb: snap.dynamic_info.memory.total / (1024 * 1024),
                         temperature_c: snap.dynamic_info.thermal.temperature.map(|t| t as f32),
                         power_watts: snap.dynamic_info.power.draw.map(|p| p as f32 / 1000.0),
-                        fan_speed_percent: snap.dynamic_info.thermal.fan_speed.map(|f| f as u8),
+                        fan_speed_percent: snap.dynamic_info.thermal.fan_speed.map(|f| f),
                         clocks: Some(GpuClocks {
                             graphics_mhz: snap.dynamic_info.clocks.graphics,
                             memory_mhz: snap.dynamic_info.clocks.memory,
